Split up client.h
