+++ google/cloud/bigquery/connection_options.cc
@@ -13,13 +13,10 @@
 // limitations under the License.
 
 #include "google/cloud/bigquery/connection_options.h"
-
+#include "google/cloud/bigquery/version.h"
 #include <grpcpp/grpcpp.h>
-
 #include <memory>
 #include <string>
-
-#include "google/cloud/bigquery/version.h"
 
 namespace google {
 namespace cloud {
