+++ google/cloud/bigquery/row.h
@@ -12,30 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BIGQUERY_CONNECTION_H_
-#define BIGQUERY_CONNECTION_H_
+#ifndef GOOGLE_CLOUD_BIGQUERY_ROW_H_
+#define GOOGLE_CLOUD_BIGQUERY_ROW_H_
 
-#include "google/cloud/bigquery/read_stream.h"
-#include "google/cloud/bigquery/row.h"
 #include "google/cloud/bigquery/version.h"
-#include "google/cloud/status_or.h"
-#include <vector>
 
 namespace google {
 namespace cloud {
 namespace bigquery {
 inline namespace BIGQUERY_CLIENT_NS {
-class Connection {
- public:
-  virtual ~Connection() = default;
+// TODO(aryann): Move all of the classes defined here except Client to their own
+// files.
 
-  virtual google::cloud::StatusOr<std::string> CreateSession(
-      std::string parent_project_id, std::string table) = 0;
-};
+// TODO(aryann): Add an implementation for a row. We must support schemas that
+// are known at compile-time as well as those that are known at run-time.
+class Row {};
 
 }  // namespace BIGQUERY_CLIENT_NS
 }  // namespace bigquery
 }  // namespace cloud
 }  // namespace google
 
-#endif  // BIGQUERY_CONNECTION_H_
+#endif  // GOOGLE_CLOUD_BIGQUERY_ROW_H_
